Fixing up subject and type issues on empty lists
List.new previously created an empty list with subject RDF.nil and
failed to replace that subject in its graph when appending the first
items. Cleared lists that were then appended to had no type (RDF.List).
Both issues are fixed in this commit, and some tests are added forsimple
list cases (i.e. not embedded in a Resource).