Make `Relation#clear` atomic
When clearing a `Relation`, we were calling `#delete` once for each
statement. By using `#delete` with a `nil` value, we can delete the
statements matching the subject/predicate pair with a single command and
count on the parent to implement an efficient `#delete`.

The default `Mutable` ends up deleting the statements individually,
anyhow, but we can probably do better. See:
https://github.com/ruby-rdf/rdf/blob/develop/lib/rdf/mixin/mutable.rb#L257-L259