Switching from Array() to Array.wrap
The Array coercion is method (i.e. Array()) aggressively attempts to
coerce the object into an Array. A Time object becomes an array of its
constituent numbers (i.e. [CCYY,MM,DD,HH,SS, TZ]). This is not the
desired effect.

Instead, it appears that we are interested in ensuring that a singular
object is wrapped in an Array. Since we are already dependent on
ActiveSupport.

Closes #178