add loaded flag to lazy load property sources with parent_strategy
The previous reload for lazy load was in Relation.make_node.  This caused a reload of all peers for a multi-value property.  The reload has moved to get_relation so that it only reloads for the individual source being referenced.

Example:  (assumes has_chapter has 5 chapters)
```
bk1.has_chapter.first.title
```

Previously, this would load all triples for all 5 chapters in has_chapter.  With the move of the reload, this only loads all triples for the first chapter.