Use a transaction when rewriting `rdf_subject`
`RDFSource#set_subject!` used to individually rewrite statements with
`#each_statement`, this meant touching every statement, even if none
would end up being rewritten. Additionally, it meant deleting and
inserting statements one at a time, even if the underlying repository
had a more efficient batch update.

This is replaced with a transaction over `#graph` and two queries.

The performance increase from this is theoretical, and not very likely
to play out in practice with the default `RDF::Repository`. This is true
for two reasons:

  - While query for the subject position will be much faster than doing
    `#each_statement` (subjects are a simple hash lookup), the `object`
    query will require a lookup on all subjects/predicate pairs. This is
    likely to have similar cost to simply iterating all the statements.

  - Bulk `#delete` and `#insert` run in linear time over the number of
    statements processed.

Still this is theoretically cleaner, and may show a benefit for other
in-memory backing repositories.