Rethink ParentStrategy based on Transaction
`#reload` calls in ParentStrategy are a major performance problem. They
spend a lot of time querying existing Graphs and copying their data into
the child's in-memory buffer Graph.

We avoid all of this by using a buffered Transaction as the the child's
Graph. The transaction is applied to `#parent` on execution, with
conflicts resolved by replaying the buffer. Reloads consist of
discarding the transaction and opening a new one against the current
parent. The cost of this comparable to a variable assignment (+
amortized GC for the old `Transaction`).