Implement a custom `Relation#<=>`
We need a hybrid comparison, using Set equality for membership, but
`#==` (like `Array`) for member comparison.

`Relation#empty?` and `Relation#length` are optimized in support of
these changes.

We need to include the extension behavior in `Relation#objects` because
the interface of `RDF::Queryable#query` return values may vary. In some
cases we can end up with an `Enumerator<RDF::Statement>` instead of an
RDF::Enumerable. There's probably a better way to handle this than the
metaprogramming approach here, but we do the same thing in at least one
place in the RDF.rb core. If there are negative side effects re: method
cache, we're already incurring them.