Remove #reload call on `RDFSource` initialization
Avoids `#reload` on `RDFSource#initialize`.

After this, we only force reloads with `ParentStrategy` and
`List`. In the former case, these calls are resyncs with the parent node
held in memory.

This causes a number of test failures that need to be evaluated. There's
probably more work to do to clean up `#initialize`, too.