Return all `<s> <p> ?o` patterns in Relation (#249)
* Return all `<s> <p> ?o` patterns in Relation

Regardless of type/class of the object, return the value. Cast to the
configured class by default. On type conflicts, cast to the specified
type.

* Relation#make_node reorganization

Reorganizes #make_node to short circuit if a cached node is
available. It used to jump through a bunch of hoops to set the node
cache to an existing value.

Q: Do we need to invalidate the relation cache if a property is changed
inline?

Fixes a few minor issues encountered in the process:

  - Casts `#build` responses regardless of `#cast?` value.
  - Sets parent properly if `ParentStrategy` is explicitly configured
  for the property.