Feature/remove auto persist (#248)
* Remove `@persisting` flag from `Persistable`

Closes #135.

This `@persisting` flag is meant to prevent looping calls to
`#persist!` via callbacks. I don't know of a use case for it, and it
seems like calling `#persist!` in a callback is a bug that should be
rooted out.

* Remove redundant `Persistable#each`

* Remove autopersist in `Relation#add_child_node`

To address #135, we're moving away from auto-persist throughout. This
call in particular is likely redundant, since `Relation#set` also
persists when a `ParentStrategy` is used.

* Limit automatic `#persist!` calls to List

* Test child node persistence behavior on #set_value